correction +de 24h et - de 30 min

+++ src/main/java/com/parkit/parkingsystem/service/FareCalculatorService.java
@@ -10,11 +10,11 @@ public class FareCalculatorService {
             throw new IllegalArgumentException("Out time provided is incorrect:"+ticket.getOutTime().toString());
         }
 
-        long inHour = ticket.getInTime().getTime()/(1000*60);
-        long outHour = ticket.getOutTime().getTime()/(1000*60);
+        double inHour = ticket.getInTime().getTime()/(1000*60);
+        double outHour = ticket.getOutTime().getTime()/(1000*60);
 
         //TODO: Some tests are failing here. Need to check if this logic is correct
-        long duration= (outHour-inHour)/60;
+        double duration= (outHour-inHour)/60;
 
 
         switch (ticket.getParkingSpot().getParkingType()){
